feat: add resource for report groups [ENG-5559] (#92)
[ENG-5559](https://stacklet.atlassian.net/browse/ENG-5559)

### what

Add reource for notification report groups.

Note that notification settings for the different transports are not yet
handled. They're present in the resource schema but not passed to the
API.
This will be added in a followup branch since this one is already quite
large.

### why

allow managing notifications via terraform

### testing

testing in sandbox, added acceptance tests

### docs

updated here


[ENG-5559]:
https://stacklet.atlassian.net/browse/ENG-5559?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ