chore: replace modelupdate with Update method on the models (#129)
### what

Move the logic from the modelupdate package into model Update() methods,
which
also incldues the rest of the update logic for the model.

Also fix a couple of bugs:
- make the `url` parameter required in Jira the configuration profile
resource, as it
 is required.
- fill in `user_fields` in the Slack configuration profile data source,
which
 was only handled in the resource

### why

Reduce duplication between resources and datasource which share a lot
similar
logic for update.

### testing

updated tests.

### docs

n/a